backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
PyJWT==2.10.1
bcrypt==4.2.1
cryptography==46.0.5
sqlalchemy==2.0.40
pydantic==2.11.1
pydantic[email]==2.11.1
pydantic-settings==2.8.1
python-dotenv==1.0.1
